Task: Move to scss, update and optimise webpack
