+++ package.json
@@ -38,7 +38,7 @@
     ],
     "moduleNameMapper": {
       "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
-      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
+      "\\.(css|scss|less)$": "<rootDir>/__mocks__/styleMock.js"
     }
   },
   "dependencies": {
@@ -92,6 +92,7 @@
     "gh-pages": "^2.0.1",
     "husky": "^1.1.3",
     "jest": "^24.7.1",
+    "mini-css-extract-plugin": "^0.8.0",
     "node-sass": "^4.12.0",
     "prettier-stylelint": "^0.4.2",
     "react-testing-library": "^5.2.3",
