+++ packages/components/media-player/src/PlayerControls/TimeBox.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import style from './index.module.css';
+import style from './index.module.scss';
 
 class TimeBox extends React.Component {
