+++ packages/components/media-player/src/PlayerControls/index.module.scss
