+++ packages/components/media-player/src/ProgressBar.module.scss
