+++ packages/components/media-player/src/config/defaultHotKeys.js
